Move SystemIteratorUtil into iteratorsImpl
